move multi user sessions
